+++ HISTORY.rst
@@ -6,6 +6,8 @@ Pending Release
 ---------------
 
 * New release notes go here.
+* Add ``PlainJSONFormatter`` to return dictionaries plain and not encoded to strings, required for ``fluent-logger``'s
+  handler.
 
 1.0.0 (2016-01-15)
 ------------------